requirements.txt
asgiref==3.2.3
astroid==2.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-bootstrap4==1.1.1
django-embed-video==1.3.1
django-precise-bbcode==1.2.12
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
isort==4.3.21
Kivy==1.11.1
kivy-deps.glew==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==5.0.1
wrapt==1.11.2
